fix(seed-demo): the contract rules the product actually enforces
The generated contracts did not survive contact with the API. Four things,
found by running it rather than by reading:

value_basis is 'total' or 'annualized_12m'. I sent "annual", and the schema
CHECK maps to the generic "the contract's dates or amounts contradict each
other" -- so the error pointed at the dates, which were fine. Reproducing it
with a body carrying no dates at all is what found the real field.

A cancellation must take effect INSIDE the term. The product refuses one that
lands after the term already ran out (contract_cancellation_within_term), and
rightly -- it would extend a term by ending it. Notice is now given part-way
through and the effect lands between notice and the end.

superseded is not assertable. draft -> superseded is refused outright; the
state exists only as the front half of a renewal. The predecessor is created
ACTIVE and renews_into supersedes it, which is the shape demo.json's own chain
uses. The successor is no longer created directly either: renewing writes its
row, and creating it first left two.

renewContract reads its successor's terms out of refs.contractsByRef, which
held only demo.json's -- a generated renewal failed with "renews_into names X,
which is not in this dataset". Each company now gets a copy with its own
contracts appended, built fresh so one company's chain cannot leak into the
next one's lookup.

Two planner bugs surfaced alongside. Coverage promotions APPEND, so a company
that kept ranking highest collected six agreements -- capContracts holds it to
three, keeping the front of the slice because a renewal chain's two halves
must survive together. And capping can drop the status just promoted, which
made fillCoverage's unbounded retry spin forever; the loop is bounded by the
company count now, and bestCandidate skips accounts too full to absorb one.

RESULT: verify passes. 101 companies across all five lifecycles, contracts in
all five statuses (23 active, 9 draft, 7 expired, 4 cancelled, 3 superseded),
leads in all four, 46 contract PDFs, 16 account documents, 7,027 facts. A
second run creates nothing.

profile.go split at 549 lines: the fill algorithm moved to coveragefill.go.
